refactor(KakaoPage): 코드 구조 개선
- WebtoonWeek enum 추가, 웹툰 구분 통일
- 쿼리 변수 타입 버그 수정
