+++ src/crawler/kakaoPage/requestAuthorsOfWebtoon.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
+import { KAKAO_PAGE_API_URL } from '.';
 
-const query = `
+const QUERY = `
 query contentHomeOverview($seriesId: Long!) {
     contentHomeOverview(seriesId: $seriesId) {
       content {
@@ -9,8 +10,6 @@ query contentHomeOverview($seriesId: Long!) {
     }
   }  
 `;
-
-const url = 'https://page.kakao.com/graphql';
 
 interface RequestResult {
   data: {
@@ -23,14 +22,14 @@ interface RequestResult {
 }
 
 export const requestAuthorsOfWebtoon = async (seriesId: string) => {
-  const { data } = await axios<RequestResult>(url, {
+  const { data } = await axios<RequestResult>({
     method: 'post',
     headers: {
       'Content-Type': 'application/json',
     },
-    url,
+    url: KAKAO_PAGE_API_URL,
     data: {
-      query,
+      QUERY,
       variables: {
         seriesId,
       },
